% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day}
\alias{day}
\title{Test day dataset}
\format{
\subsection{\code{day}}{

A data frame with 5,602 rows and 13 columns:
\describe{
\item{hh_id}{8 digit household ID}
\item{person_id}{10 digit person ID}
\item{day_id}{12 digit day ID}
\item{delivery_2}{Take-out/prepared food delivered to home, response codes}
\item{delivery_3}{Someone came to do work at home (e.g., babysitter,
housecleaning, lawn), response codes}
\item{delivery_4}{Groceries delivered to home, response codes}
\item{delivery_5}{Received packages at home (e.g., USPS, FedEx, UPS), response codes}
\item{delivery_6}{Received personal packages at work, response codes}
\item{delivery_7}{Received packages at another location (e.g.,
Amazon Locker, package pick-up point), response codes}
\item{delivery_8}{Other item delivered to home (e.g., appliance), response codes}
\item{delivery_996}{None of the above, response codes}
\item{begin_day}{Location at the beginning of the day, response codes}
\item{end_day}{Location at the end of the day, response codes}
}
}
}
\source{
Export tables compiled from internal RSG database on 11/22/2023
}
\usage{
day
}
\description{
A dataset containing information about participant travel days from the
2023 Puget Sound Regional Council Household Travel Survey.
The variables are as follows:
}
\keyword{datasets}
