% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize_df.R
\name{factorize_df}
\alias{factorize_df}
\title{Factorize a dataframe}
\usage{
factorize_df(df, vals_df, verbose = TRUE, ...)
}
\arguments{
\item{df}{A dataframe to label}

\item{vals_df}{A dataframe of variable labels (i.e., factor levels and
labels) with the format as specified below. Passed to factorize_column
function.}

\item{verbose}{Prints which vars are labeled and unlabeled}

\item{...}{Additional arguments passed to \code{\link{factorize_column}}}
}
\value{
A factorized (i.e. labeled) version of the dataframe it was passed.
}
\description{
Factorize a dataframe. The function loops over a dataframe (calling
factorize_column) and labels each variable for which you provide labels.
}
\details{
The function expects a values dataframe (vals_df) in the following format:
variable (the character/string names of each variable), value (the integer
values for each variable), val_order (the sequential ordering of each
value), label (the strings or names to use in the levels of the factor).

The "factorize" functions were borrowed and updated from the
'tmr.Rite.out.tester' package by Matt Landis.
}
\examples{

hh_labeled = factorize_df(
  df = hh,
  vals_df = value_labels,
  value_label_colname = "label",
  extra_labels = c("Missing")
)

}
