% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance_meters.R
\name{get_distance_meters}
\alias{get_distance_meters}
\title{Get distance between points}
\usage{
get_distance_meters(location_1, location_2, radius = 6378137)
}
\arguments{
\item{location_1}{A matrix or vector of longitudes and latitudes}

\item{location_2}{A matrix or vector of longitudes and latitudes}

\item{radius}{Radius of the sphere to use for haversine calculation (Defaults to meters)}
}
\value{
A vector of distances in meters
}
\description{
Function to get haversine distance in meters between two points
Based on calculation from the geosphere package
}
