% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hh}
\alias{hh}
\title{Test household travel survey dataset}
\format{
\subsection{\code{test_data}}{

A list containing the following datasets:

Test household dataset

A dataset containing household level attributes of 1,000 households from the
2023 Puget Sound Regional Council Household Travel Survey.
The variables are as follows:
}

\subsection{\code{hh}}{

A data frame with 1,000 rows and 6 columns:
\describe{
\item{hh_id}{8 digit household ID}
\item{sample_segment}{Sample segment}
\item{income_detailed}{2022 household income (detailed categories), response codes}
\item{income_followup}{2022 household income (broad categories), response codes}
\item{num_people}{Household size, response codes}
\item{residence_type}{Type of residence, response codes}
}
}
}
\source{
Export tables compiled from internal RSG database on 11/22/2023
}
\usage{
hh
}
\description{
Test household travel survey dataset
}
\keyword{datasets}
