% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_bin_var.R
\name{hts_bin_var}
\alias{hts_bin_var}
\title{Bin numeric variables}
\usage{
hts_bin_var(prepped_dt, numvar, nbins = 7)
}
\arguments{
\item{prepped_dt}{Dataset containing variable to bin in data.table format}

\item{numvar}{Name of the numeric variable to bin}

\item{nbins}{Number of bins for variable. Defaults to 7.}
}
\value{
Inputted dataset with the specified variable binned in data.table format.
}
\description{
Bin numeric variables
}
\examples{

require(data.table)
require(stringr)
hts_bin_var(prepped_dt = trip, numvar = 'speed_mph')

}
