% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_calculate_vmt.R
\name{hts_calculate_vmt}
\alias{hts_calculate_vmt}
\title{Calculate vmt for each trip}
\usage{
hts_calculate_vmt(
  trip_dt,
  mode_cols,
  miles_col,
  vehicle_modes,
  occupancy_var = NULL
)
}
\arguments{
\item{trip_dt}{Data table of trip data}

\item{mode_cols}{Column(s) in trip_dt containing trip mode}

\item{miles_col}{Column in trip_dt containing miles per trip}

\item{vehicle_modes}{List of modes that are considered vehicle}

\item{occupancy_var}{Ocuupancy column to divide distance by if specified. Default is NULL}
}
\value{
List of vmt for each trip in trip_dt
}
\description{
Calculate vmt for each trip
}
\examples{

test_data$trip[, vmt := 
 hts_calculate_vmt(
  trip_dt = test_data$trip,
  mode_cols = c('mode_1', 'mode_2'),
  miles_col = 'distance_miles',
  vehicle_modes = c(6, 7, 10)
 )
]
test_data$trip[, vmt := 
 hts_calculate_vmt(
  trip_dt = test_data$trip,
  mode_cols = 'mode_type',
  miles_col = 'distance_miles',
  vehicle_modes = 8,
  occupancy_var = 'num_travelers'
 )
]
}
