% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_cbind_var.R
\name{hts_cbind_var}
\alias{hts_cbind_var}
\title{Bind a column from one table to another}
\usage{
hts_cbind_var(
  lhs_table,
  rhs_var,
  variable_list = variable_list,
  return_weight_cols = FALSE,
  ...
)
}
\arguments{
\item{lhs_table}{Table to bind a column to in data.table format}

\item{rhs_var}{Variable to bind to the lhs_table.}

\item{variable_list}{A variable list with descriptions and table locations
of variables.}

\item{return_weight_cols}{If true binds weight variable along with rhs_var
to lhs_table. Default is FALSE.}

\item{...}{Additional arguments passed to \code{link{hts_get_keycols}}}
}
\value{
Inputted table with inputted variable binded.
}
\description{
Bind a column from one table to another
}
\examples{

require(data.table)
hts_cbind_var(lhs_table = trip, rhs_var = 'speed_mph', variable_list = variable_list)
hts_cbind_var(lhs_table = trip, rhs_var = 'speed_mph', 
variable_list = variable_list, return_weight_cols = TRUE)

}
