% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_filter_data.R
\name{hts_filter_data}
\alias{hts_filter_data}
\title{Filter datasets to only keep specified ids}
\usage{
hts_filter_data(hts_data, ids, id_name)
}
\arguments{
\item{hts_data}{List of containing household travel data tables.}

\item{ids}{List of ids to keep in all of the tables}

\item{id_name}{Name of id being used for filtering (e.g., hh_id, person_id)}
}
\value{
Inputted list of tables filtered to the specified ids.
}
\description{
Filter datasets to only keep specified ids
}
\examples{

require(data.table)

data(test_data)

hts_filter_data(
  hts_data = test_data,
  ids = hh[num_people > 5, hh_id],
  id_name = "hh_id"
)

}
