% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_find_var.R
\name{hts_find_var}
\alias{hts_find_var}
\title{Find table location of a variable}
\usage{
hts_find_var(var, variables_dt = variable_list)
}
\arguments{
\item{var}{searchable variable in string format}

\item{variables_dt}{codebook variable list in data table format}
}
\value{
variable location in string format
}
\description{
Find table location of a variable
}
\examples{
var_location = hts_find_var('income_detailed')

}
