% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_get_keycols.R
\name{hts_get_keycols}
\alias{hts_get_keycols}
\title{Find key columns in table}
\usage{
hts_get_keycols(dt, ids = TRUE, weights = TRUE, priority = FALSE)
}
\arguments{
\item{dt}{Dataset to find key columns of in data.table format}

\item{ids}{Boolean whether to return id columns. Default is TRUE.}

\item{weights}{Boolean whether to return weight columns. Default is TRUE.}

\item{priority}{Boolean whether to only return highest level weight/id.
Default is FALSE.}
}
\value{
List of names of key columns in the dataset.
}
\description{
Find key columns in table
}
\examples{

require(data.table)
hts_get_keycols(dt = trip)
hts_get_keycols(dt = trip, priority = TRUE)

}
