% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_melt_vars.R
\name{hts_melt_vars}
\alias{hts_melt_vars}
\title{Melts checkbox variables into a single variable}
\usage{
hts_melt_vars(
  shared_name = NULL,
  wide_dt = NULL,
  shared_name_vars = NULL,
  variables_dt = variable_list,
  data = hts_data,
  ids = c("hh_id", "person_id", "day_id", "trip_id", "vehicle_id"),
  remove_missing = TRUE,
  missing_values = c("Missing Response", "995"),
  checkbox_label_sep = ":",
  to_single_row = FALSE
)
}
\arguments{
\item{shared_name}{Shared name of the checkbox variable to melt (e.g., 'race_1',
and 'race_2' have shared name of 'race'). Defaults to NULL.}

\item{wide_dt}{Table containing the checkbox variables to melt in data.table
format.}

\item{shared_name_vars}{List of the checkbox variables with the same shared name.
Defaults to NULL.}

\item{variables_dt}{List of variable locations and descriptions in data.table
format.}

\item{data}{List containing household, person, day, trip, and vehicle
datasets in data.table format.}

\item{ids}{unique identifiers appearing in wide_dt}

\item{remove_missing}{Boolean to remove rows with missing values. Defaults to
TRUE.}

\item{missing_values}{Missing values to remove. Defaults to 'Missing Response'
and 995.}

\item{checkbox_label_sep}{Character to use to split description of checkbox
variable in the variable list. Defaults to ':'}

\item{to_single_row}{Boolean if treating multiple checkbox selections as 'Two
or more'. Defaults to FALSE.}
}
\value{
Inputted data table with checkbox variables melted into a single variable,
with a 'variable' column to indicate original variable names and a 'value' column
with the original value of the checkbox variable.
}
\description{
Melts checkbox variables into a single variable
}
\examples{

require(data.table)
require(stringr)
hts_melt_vars(
  shared_name = "race",
  wide_dt = person,
  data = list(
    "hh" = hh,
    "person" = person,
    "day" = day,
    "trip" = trip,
    "vehicle" = vehicle
  )
)

}
