% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_prep_byvar.R
\name{hts_prep_byvar}
\alias{hts_prep_byvar}
\title{Prepare variable to be summarized}
\usage{
hts_prep_byvar(
  summarize_by = NULL,
  variables_dt = variables_list,
  hts_data,
  ...
)
}
\arguments{
\item{summarize_by}{Name of the variable to be prepped. Default is NULL.}

\item{variables_dt}{List of variable locations and descriptions in data.table
format.}

\item{hts_data}{List containing household, person, day, trip, and vehicle
datasets in data.table format.}

\item{...}{Additional parameters to pass to \code{link{hts_melt_vars}}}
}
\value{
Data table containing the variable to be summarized and other key
columns.
}
\description{
Prepare variable to be summarized
}
\examples{

hts_prep_byvar(summarize_by = 'age',
               variables_dt = variable_list,
               hts_data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
hts_prep_byvar(summarize_by = 'race',
               variables_dt = variable_list,
               hts_data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))

}
