% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_prep_data.R
\name{hts_prep_data}
\alias{hts_prep_data}
\title{Prepare datasets to make summaries}
\usage{
hts_prep_data(
  summarize_var = NULL,
  summarize_by = NULL,
  variables_dt = variable_list,
  data = hts_data,
  id_cols = NULL,
  wt_cols = NULL,
  weighted = TRUE,
  remove_outliers = TRUE,
  threshold = 0.975,
  remove_missing = TRUE,
  missing_value = 995,
  not_imputable = -1,
  strataname = NULL
)
}
\arguments{
\item{summarize_var}{Name of the variable to summarize. Default is NULL}

\item{summarize_by}{Name of the variable to summarize the summarize_var by.
Default is NULL.}

\item{variables_dt}{List of variable locations and descriptions in data.table
format.}

\item{data}{List of household, person, vehicle, day, and trip tables in
data.table format.}

\item{weighted}{Whether the data is weighted. Default is TRUE.}

\item{remove_outliers}{Whether to remove outliers for numeric variable. Default
is TRUE.}

\item{threshold}{Threshold to define outliers. Default is 0.975.}

\item{remove_missing}{Whether to remove missing values from the summary.
Default is TRUE.}

\item{missing_value}{Missing value to remove. Default is 995.}

\item{not_imputable}{Value representing 'Not imputable' to remove. Default
is -1.}

\item{strataname}{Name of strata name to bring in. Default is NULL.}
}
\value{
List containing the categorical and numeric datasets of the summary
variables and key columns, and either whether the summarize variable is shared
or a breakdown of outliers, depending on if the summarize variable is
categorical or numeric.
}
\description{
Prepare datasets to make summaries
}
\examples{

require(data.table)
require(stringr)
hts_prep_data(summarize_var = 'age',
              variables_dt = variable_list,
              data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
hts_prep_data(summarize_var = 'speed_mph',
              summarize_by = 'age',
              variables_dt = variable_list,
              data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
                            
                            
hts_prep_data(summarize_var = 'employment',
              summarize_by = c('age', 'race'),
              variables_dt = variable_list,
              data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
}
