% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_prep_vmtrate.R
\name{hts_prep_vmtrate}
\alias{hts_prep_vmtrate}
\title{Prepare datasets for VMT rate calculations}
\usage{
hts_prep_vmtrate(
  summarize_by = NULL,
  variables_dt = variable_list,
  trip_name = "trip",
  day_name = "day",
  ids = c("hh_id", "person_id", "day_id", "trip_id", "vehicle_id"),
  traveler_count_var,
  dist_var,
  mode_var,
  veh_regex,
  wts = c("hh_weight", "person_weight", "day_weight", "trip_weight", "hh_weight"),
  remove_outliers = FALSE,
  threshold = 0.975,
  weighted = TRUE,
  strataname = NULL,
  hts_data = list(hh = hh, person = person, day = day, trip = trip, vehicle = vehicle)
)
}
\arguments{
\item{summarize_by}{Name of the variable to summarize vmt rates by. Default
is NULL.}

\item{variables_dt}{List of variable locations and descriptions in data.table
format.}

\item{trip_name}{Name of the trip dataset in hts_data.}

\item{day_name}{Name of the day dataset in hts_data.}

\item{ids}{name of unique identifier in each table in hts_data}

\item{traveler_count_var}{name of traveler count variable in trip table}

\item{dist_var}{name of distance variable in trip table, preferably miles for VMT}

\item{mode_var}{name of mode variable in trip table}

\item{veh_regex}{regular expression pattern capturing vehicular modes in mode_var}

\item{wts}{name of weight column in each table in hts_data}

\item{remove_outliers}{Boolean whether or not to remove outliers from dataset.
Default is TRUE.}

\item{threshold}{Threshold to define outliers. Default is 0.975.}

\item{weighted}{Whether the data is weighted. Default is TRUE.}

\item{strataname}{Name of strata name to bring in. Default is NULL.}

\item{hts_data}{List containing household, person, day, trip, and vehicle
datasets in data.table format.}
}
\value{
List of binned vehicle miles traveled with key columns and summarize by variable,
unbinned vehicle miles traveled with key columns and summarize by variable, and a
breakdown of outliers if removed.
}
\description{
Prepare datasets for VMT rate calculations
}
