% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_remove_missing_data.R
\name{hts_remove_missing_data}
\alias{hts_remove_missing_data}
\title{Remove missing data for summary variables}
\usage{
hts_remove_missing_data(
  hts_data,
  variables_dt,
  summarize_var,
  ids = NULL,
  summarize_by = NULL,
  missing_value = 995,
  not_imputable = -1
)
}
\arguments{
\item{hts_data}{List containing household, person, day, trip, and vehicle
datasets in data.table format.}

\item{variables_dt}{A variable list with descriptions and table locations
of variables.}

\item{summarize_var}{Variable to be summarized that has it's missing data
removed.}

\item{summarize_by}{Variable being summarized by that has it's missing data
removed. Default is NULL.}

\item{missing_value}{Missing value that will be removed. Default is 995.}

\item{not_imputable}{Value meaning not_imputable that will be removed. Default
is -1.}
}
\value{
Inputted list of datasets without missing values for specified variables.
}
\description{
Remove missing data for summary variables
}
\examples{

require(data.table)
hts_remove_missing_data(hts_data = list('hh' = hh,
                                        'person' = person,
                                        'day' = day,
                                        'trip' = trip,
                                        'vehicle' = vehicle),
                        variables_dt = variable_list,
                        summarize_var = 'speed_mph',
                        summarize_by = 'mode_type')

}
