% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_remove_outliers.R
\name{hts_remove_outliers}
\alias{hts_remove_outliers}
\title{Remove outliers from a numeric variable}
\usage{
hts_remove_outliers(var_dt, numvar = NULL, threshold = 0.975)
}
\arguments{
\item{var_dt}{Dataset with a numeric variable to remove outliers from
in data.table format.}

\item{numvar}{Numeric variable to remove outliers from. Default is NULL.}

\item{threshold}{Threshold to define what an outlier is. Default is .975.}
}
\value{
List of outliers removed and the dataset without the outliers.
}
\description{
Remove outliers from a numeric variable
}
\examples{

require(data.table)
hts_remove_outliers(var_dt = hh, numvar = 'num_people')

}
