% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_summary.R
\name{hts_summary}
\alias{hts_summary}
\title{Title}
\usage{
hts_summary(
  hts_data = list(hh = hh, person = person, day = day, trip = trip, vehicle = vehicle,
    location = location),
  summarize_var = NULL,
  summarize_by = NULL,
  variables_dt = variable_list,
  var_description_sep = ":",
  values_dt = value_labels,
  weighted = TRUE,
  wtname = NULL,
  strataname = NULL,
  remove_outliers = TRUE,
  se = FALSE,
  threshold = 0.975,
  remove_missing = TRUE,
  missing_value = 995,
  not_imputable = -1
)
}
\arguments{
\item{not_imputable}{}
}
\description{
Title
}
