% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_summary_cat.R
\name{hts_summary_cat}
\alias{hts_summary_cat}
\title{Summarize a categorical variable}
\usage{
hts_summary_cat(
  prepped_dt,
  summarize_var = NULL,
  summarize_by = NULL,
  weighted = TRUE,
  se = FALSE,
  wtname = NULL,
  strataname = NULL,
  variable_list = NULL,
  values_dt = value_labels,
  extra_labels = extra_labels
)
}
\arguments{
\item{prepped_dt}{Dataset containing the summary variables and key columns in
data.table format.}

\item{summarize_var}{Name of the categorical variable to summarize. Default is NULL.}

\item{summarize_by}{Name of the variable to summarize the summarize_var by.
Default is NULL.}

\item{weighted}{Whether the data is weighted. Default is TRUE.}

\item{se}{Whether to calculate standard error. Default is FALSE.}

\item{wtname}{Name of the weight column to use. Default is NULL.}

\item{strataname}{Name of strata name to bring in. Default is NULL.}

\item{variable_list}{Dataset of variable locations and descriptions in
data.table format.}

\item{values_dt}{Dataset of values and value labels for all variables in
data.table format.}

\item{extra_labels}{Pass a vector of the names of other variables in the
values dataframe to use when labeling. Common uses include missing values,
universal values (e.g., "Total" or "Subtotal" row), or similar variables
(e.g., "mode_1" to reuse the mode_1 labels for mode_2). Default is NULL.}
}
\value{
List of unweighted and weighted categorical summaries including counts
and proportions.
}
\description{
Summarize a categorical variable
}
\examples{

require(data.table)
require(stringr)
require(dplyr)
require(srvyr)
DT = hts_prep_data(summarize_var = 'age',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$cat
hts_summary_cat(prepped_dt = DT,
                summarize_var = 'age',
                values_dt = value_labels)
DT = hts_prep_data(summarize_var = 'age',
                   summarize_by = 'employment',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$cat
hts_summary_cat(prepped_dt = DT,
                summarize_var = 'age',
                summarize_by = 'employment',
                values_dt = value_labels,
                extra_labels = 'Missing')
}
