% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_summary_num.R
\name{hts_summary_num}
\alias{hts_summary_num}
\title{Summarize a numeric variable}
\usage{
hts_summary_num(
  prepped_dt,
  summarize_var = NULL,
  summarize_by = NULL,
  weighted = TRUE,
  values_dt = value_labels,
  wtname = NULL,
  strataname = NULL
)
}
\arguments{
\item{prepped_dt}{Dataset containing the summary variables and key columns in
data.table format.}

\item{summarize_var}{Name of the numeric variable to summarize. Default is NULL.}

\item{summarize_by}{Name of the variable to summarize the summarize_var by.
Default is NULL.}

\item{weighted}{Whether the data is weighted. Default is TRUE.}

\item{values_dt}{Dataset of values and value labels for all variables in
data.table format.}

\item{wtname}{Name of the weight column to use. Default is NULL.}

\item{strataname}{Name of strata name to bring in. Default is NULL.}
}
\value{
List of unweighted and weighted numeric summaries including count, min,
max, mean, se, and median.
}
\description{
Summarize a numeric variable
}
\examples{

require(data.table)
require(stringr)
require(dplyr)
require(srvyr)
DT = hts_prep_data(summarize_var = 'speed_mph',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$num
hts_summary_num(prepped_dt = DT,
                summarize_var = 'speed_mph',
                values_dt = value_labels)
DT = hts_prep_data(summarize_var = 'speed_mph',
                   summarize_by = 'age', 
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$num
hts_summary_num(prepped_dt = DT,
                summarize_var = 'speed_mph',
                summarize_by = 'age',
                values_dt = value_labels)
}
