% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_summary_wrapper.R
\name{hts_summary_wrapper}
\alias{hts_summary_wrapper}
\title{Make household travel survey summaries-- runs hts_prep_variable and hts_summary}
\usage{
hts_summary_wrapper(
  summarize_var = NULL,
  summarize_by = NULL,
  variables_dt = variable_list,
  vals_df = value_labels,
  data = list(hh = hh, person = person, day = day, trip = trip, vehicle = vehicle),
  id_cols = c("hh_id", "person_id", "day_id", "trip_id", "vehicle_id"),
  weighted = TRUE,
  wt_cols = c("hh_weight", "person_weight", "day_weight", "trip_weight", "hh_weight"),
  trip_name = "trip",
  day_name = "day",
  strataname = NULL,
  se = FALSE,
  checkbox_valname = "value",
  checkbox_yesval = 1,
  value_label_colname = "label",
  remove_outliers = TRUE,
  threshold = 0.975,
  remove_missing = TRUE,
  not_imputable = -1,
  missing_values = c("Missing Response", "995")
)
}
\arguments{
\item{summarize_var}{Name of the variable to summarize. Default is NULL}

\item{summarize_by}{Name of the variable to summarize the summarize_var by.
Default is NULL.}

\item{variables_dt}{List of variable locations and descriptions in data.table
format.}

\item{vals_df}{A dataframe of variable labels (i.e., factor levels and
labels) with the format as specified below. Passed to factorize_column
function.}

\item{data}{List of household, person, vehicle, day, and trip tables in
data.table format.}

\item{id_cols}{name of unique identifier for each table in hts_data}

\item{weighted}{Whether the data is weighted. Default is TRUE.}

\item{wt_cols}{weight name for each table in hts_data}

\item{trip_name}{Name of the trip dataset in hts_data.}

\item{day_name}{Name of the day dataset in hts_data.}

\item{strataname}{Name of strata name to bring in. Default is NULL.}

\item{se}{Whether to calculate standard error. Default is FALSE. Will be set
to FALSE if weighted is FALSE.}

\item{checkbox_valname}{Name of the column with the checkbox value. Default is 'value'.
Must be provided if summarize_var is a checkbox variable.}

\item{checkbox_yesval}{Value of checkbox_valname that indicates it was selected.
Default is 1. Must be provided if summarize_var is a checkbox variable.}

\item{value_label_colname}{The name of the value label column in vals_df. Default is 'value_label'}

\item{remove_outliers}{Whether to remove outliers for numeric variable. Default
is TRUE.}

\item{threshold}{Threshold to define outliers. Default is 0.975.}

\item{remove_missing}{Whether to remove missing values from the summary.
Default is TRUE.}

\item{not_imputable}{Value representing 'Not imputable' to remove. Default
is -1.}

\item{missing_values}{Missing values to remove. Default is 995.}
}
\value{
A list containing (if applicable) categorical and numeric summaries of the
specified variable(s), as well as sample sizes and whether or not the summarized
variable is a shared checkbox variable.
To access the categorical/numeric df use output$summary.
To access the weighted df use output$summary$wtd, and output$summary$unwtd for the
unweighted df.
To access the weight name use output$summary$weight_name.
To access sample sizes use output$n_ls.
To access weighted and unweighted sample sizes respectively, use output$n_ls$wtd
and output$n_ls$unwtd.
}
\description{
Make household travel survey summaries-- runs hts_prep_variable and hts_summary
}
\examples{


hts_summary_wrapper(
summarize_var = 'employment',
summarize_by = 'income_detailed')


hts_summary_wrapper(
summarize_var = 'race',
summarize_by = c('age', 'employment')
)

hts_summary_wrapper(
summarize_var = 'num_trips',
summarize_by = 'age')


}
