% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_trip_vehid.R
\name{hts_trip_vehid}
\alias{hts_trip_vehid}
\title{Add vehicle_id to trip table}
\usage{
hts_trip_vehid(
  trip_table,
  vehicle_table,
  vehicle_mode_type = "Vehicle",
  values_dt = value_labels,
  ...
)
}
\arguments{
\item{trip_table}{Dataset of trips in data.table format.}

\item{vehicle_table}{Dataset of vehicles in data.table format.}

\item{vehicle_mode_type}{Mode type label for vehicle. Default is 'Vehicle'.}

\item{values_dt}{Dataset of value labels in data.table format.}

\item{...}{Additional arguments passed to \code{link{factorize_column}}}
}
\value{
Trip table with vehicle_id attached.
}
\description{
Add vehicle_id to trip table
}
\examples{

require(data.table)
trip_ex = data.table(
      hh_id = sample(1:10, size = 30, replace = TRUE),
      trip_id = 1:30,
      mode_type = sample(1:2, size = 30, replace = TRUE),
      mode_1 = sample(1, size = 30, replace = TRUE))
vehicle_ex = data.table(
      hh_id = sample(1:10, size = 30, replace = TRUE),
      vehicle_id = 1:30)
values_ex = data.table(
      variable = c(rep('mode_type', 2), ('mode_1')),
      value = c(1,2,1),
      value_label = c('Vehicle', 'Walk', 'Car')
      )
hts_trip_vehid(trip_table = trip_ex,
               vehicle_table = vehicle_ex,
               vehicle_mode_type = 'Vehicle',
               values_dt = values_ex)

}
