% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{person}
\alias{person}
\title{Test person dataset}
\format{
\subsection{\code{person}}{

A data frame with 1,999 rows and 19 columns:
\describe{
\item{hh_id}{8 digit household ID}
\item{person_id}{10 digit person ID}
\item{ethnicity_1}{Not of Hispnic, Latino, or Spanish origin, response codes}
\item{ethnicity_2}{Mexican, Mexican American, Chicano, response codes}
\item{ethnicity_3}{Puerto Rican, response codes}
\item{ethnicity_4}{Cuban, response codes}
\item{ethnicity_997}{Another Hispanic, Latino, or Spanish origin, response codes}
\item{ethnicity_999}{Prefer not to answer, response codes}
\item{race_1}{African American or Black, response codes}
\item{race_2}{Americian Indian or Alaska Native, response codes}
\item{race_3}{Asian, response codes}
\item{race_4}{Native Hawaiian or other Pacific Islander, response codes}
\item{race_5}{White, response codes}
\item{race_997}{Other race, response codes}
\item{race_999}{Prefer not to answer, response codes}
\item{age}{Age, response codes}
\item{gender}{Gender, response codes}
\item{employment}{Employment status, response codes}
\item{education}{Highest level of education completed, response codes}
}
}
}
\source{
Export tables compiled from internal RSG database on 11/22/2023
}
\usage{
person
}
\description{
A dataset containing person level attributes of almost 2,000 persons from the
2023 Puget Sound Regional Council Household Travel Survey.
The variables are as follows:
}
\keyword{datasets}
