% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test household travel survey dataset}
\format{
\subsection{\code{test_data}}{

A list containing the following datasets:
}
}
\source{
A list of export tables compiled from internal RSG database on 11/22/2023
}
\usage{
test_data
}
\description{
Test household travel survey dataset
}
\keyword{datasets}
