% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trip}
\alias{trip}
\title{Test trip dataset}
\format{
\subsection{\code{trip}}{

A data frame with 21,378 rows and 7 columns:
\describe{
\item{hh_id}{8 digit household ID}
\item{person_id}{10 digit person ID}
\item{day_id}{12 digit day ID}
\item{trip_id}{13 digit trip ID}
\item{travel_date}{Date of trip}
\item{mode_type}{Type of transportation used for trip, response codes}
\item{d_purpose_category}{Purpose for taking trip to destination, response codes}
}
}
}
\source{
Export tables compiled from internal RSG database on 11/22/2023
}
\usage{
trip
}
\description{
A dataset containing attributes of over 20,000 trips from the
2023 Puget Sound Regional Council Household Travel Survey.
The variables are as follows:
}
\keyword{datasets}
