% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{value_labels}
\alias{value_labels}
\title{List of values and their labels}
\format{
\subsection{\code{value_labels}}{

A data frame with 210 rows and 3 columns:
\describe{
\item{variable}{Name of the variable}
\item{value}{The numeric value of the variable}
\item{label}{What the numeric value of the variable represents}
}
}
}
\usage{
value_labels
}
\description{
A dataset containing the values for all variables found in variable_list
The variables are as follows:
}
\keyword{datasets}
