% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variable_list}
\alias{variable_list}
\title{List of variables}
\format{
\subsection{\code{variable_list}}{

A data frame with 41 rows and 13 columns:
\describe{
\item{order}{The order the variables are presented in}
\item{source}{Where the variable was created}
\item{variable}{Name of the variable}
\item{is_checkbox}{The variable is a 'Select all that Apply' question}
\item{hh}{The variable exists in the hh table}
\item{person}{The variable exists in the person table}
\item{day}{The variable exists in the day table}
\item{trip}{The variable exists in the trip table}
\item{vehicle}{The variable exists in the vehicle table}
\item{data_type}{Data type of the variable}
\item{description}{A description of the variable}
\item{logic}{Conditions where the variable should have a value}
}
}
}
\usage{
variable_list
}
\description{
A dataset containing information about all variables existing in the hh, person,
day, trip, and vehicle tables. The variables are as follows:
}
\keyword{datasets}
