% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vehicle}
\alias{vehicle}
\title{Test vehicle dataset}
\format{
\subsection{\code{vehicle}}{

A data frame with 1,366 rows and 3 columns:
\describe{
\item{hh_id}{8 digit household ID}
\item{vehicle_id}{10 digit vehicle ID}
\item{fuel_type}{Type of fuel taken by vehicle, response codes}
}
}
}
\source{
Export tables compiled from internal RSG database on 11/22/2023
}
\usage{
vehicle
}
\description{
A dataset containing attributes of over 6,500 vehicles from the
2023 Puget Sound Regional Council Household Travel Survey.
The variables are as follows:
}
\keyword{datasets}
