% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_calculate_vmt.R
\name{hts_calculate_vmt}
\alias{hts_calculate_vmt}
\title{Calculate vmt for each trip}
\usage{
hts_calculate_vmt(
  data,
  trip_name = "trip",
  ids = c("hh_id", "person_id", "day_id", "trip_id", "vehicle_id"),
  agg_tbl = "trip",
  mode_cols,
  miles_col,
  vehicle_modes,
  occupancy_var = NULL
)
}
\arguments{
\item{data}{List of data tables}

\item{trip_name}{Name of trip table in data}

\item{ids}{Unique id in order for each table in data}

\item{agg_tbl}{Table to append vmt to}

\item{mode_cols}{Column(s) in trip_dt containing trip mode}

\item{miles_col}{Column in trip_dt containing miles per trip}

\item{vehicle_modes}{List of modes that are considered vehicle}

\item{occupancy_var}{Ocuupancy column to divide distance by if specified. Default is NULL}
}
\value{
List of vmt for each trip in trip_dt
}
\description{
Calculate vmt for each trip
}
\examples{

hts_calculate_vmt(
  trip_name = "trip",
  data = test_data,
  agg_tbl = "day",
  mode_cols = c("mode_1", "mode_2"),
  miles_col = "distance_miles",
  vehicle_modes = c(6, 7, 10)
)


hts_calculate_vmt(
  data = test_data,
  trip_name = "trip",
  agg_tbl = "trip",
  mode_cols = "mode_type",
  miles_col = "distance_miles",
  vehicle_modes = 8,
  occupancy_var = "num_travelers"
)
}
