% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_filter_data.R
\name{hts_filter_data}
\alias{hts_filter_data}
\title{Filter datasets to only keep specified ids}
\usage{
hts_filter_data(hts_data, ids, id_type = "hh")
}
\arguments{
\item{hts_data}{List of containing household, person, day, trip, and vehicle
tables in data.table format}

\item{ids}{List of ids to keep in all of the tables}

\item{id_type}{Type of id being used for filtering. Options are 'hh', 'person',
'day', and 'trip'. Defaults to 'hh'.}
}
\value{
Inputted list of tables filtered to the specified ids.
}
\description{
Filter datasets to only keep specified ids
}
\examples{

require(data.table)
hts_filter_data(hts_data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle),
                ids = hh[num_people > 5, hh_id],
                id_type = 'hh')

}
