% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_find_var.R
\name{hts_find_var}
\alias{hts_find_var}
\title{Find table location of a variable}
\usage{
hts_find_var(var, data, variables_dt = variable_list)
}
\arguments{
\item{var}{searchable variable in string format}

\item{data}{List of household, person, vehicle, day, and trip tables in
data.table format.}

\item{variables_dt}{codebook variable list in data table format}
}
\value{
variable location in string format
}
\description{
Find table location of a variable
}
\examples{

hts_find_var("income_detailed", data = test_data)

}
