% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_get_ns.R
\name{hts_get_ns}
\alias{hts_get_ns}
\title{Get counts from dataset}
\usage{
hts_get_ns(prepped_dt, weighted, ids, wt_col)
}
\arguments{
\item{prepped_dt}{Dataset to pull counts from.}

\item{weighted}{Boolean whether to pull weighted estimates.}
}
\value{
List of unweighted counts, weighted counts, and highest level unit.
}
\description{
Get counts from dataset
}
\examples{

require(data.table)
hts_get_ns(prepped_dt = day, weighted = TRUE)

}
