% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_summary_cat.R
\name{hts_summary_cat}
\alias{hts_summary_cat}
\title{Summarize a categorical variable}
\usage{
hts_summary_cat(
  prepped_dt,
  summarize_var = NULL,
  summarize_by = NULL,
  weighted = TRUE,
  se = FALSE,
  wtname = NULL,
  strataname = NULL,
  checkbox_valname = "value",
  checkbox_yesval = 1,
  summarize_vartype = "categorical"
)
}
\arguments{
\item{prepped_dt}{Dataset containing the summary variables and key columns in
data.table format.}

\item{summarize_var}{Name of the categorical variable to summarize. Default is NULL.}

\item{summarize_by}{Name of the variable to summarize the summarize_var by.
Default is NULL.}

\item{weighted}{Whether the data is weighted. Default is TRUE.}

\item{se}{Whether to calculate standard error. Default is FALSE.}

\item{wtname}{Name of the weight column to use. Default is NULL.}

\item{strataname}{Name of strata name to bring in. Default is NULL.}

\item{checkbox_valname}{Name of the column with the checkbox value. Default is 'value'.
Must be provided if summarize_var is a checkbox variable.}

\item{checkbox_yesval}{Value of checkbox_valname that indicates it was selected.
Default is 1. Must be provided if summarize_var is a checkbox variable.}

\item{summarize_vartype}{String; one of either 'categorical' (when the
variable being summarized is categorical) or 'checkbox' (when the variable being
summarized is derived from a multiple response, aka select-all-that-apply question).}
}
\value{
List of unweighted and weighted categorical summaries including counts
and proportions.
}
\description{
Summarize a categorical variable
}
\examples{

require(data.table)
require(stringr)
require(dplyr)
require(srvyr)
DT = hts_prep_data(summarize_var = 'age',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$cat
hts_summary_cat(prepped_dt = DT,
                summarize_var = 'age',
                wtname = 'person_weight')
                
DT = hts_prep_data(summarize_var = 'race',
                   summarize_by = 'employment',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$cat
hts_summary_cat(prepped_dt = DT,
                summarize_var = 'race',
                summarize_by = 'employment',
                wtname = 'person_weight',
                summarize_vartype = 'checkbox')
                
DT = hts_prep_data(summarize_var = 'employment',
                summarize_by = c('race', 'income_detailed', 'gender'),
                variables_dt = variable_list,
                data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))$cat

hts_summary_cat(prepped_dt = DT,
                summarize_var = 'employment',
                summarize_by = c('race', 'income_detailed', 'gender'),
                wtname = 'person_weight')$unwtd
}
