% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts_summary_num.R
\name{hts_summary_num}
\alias{hts_summary_num}
\title{Summarize a numeric variable}
\usage{
hts_summary_num(
  prepped_dt,
  summarize_var = NULL,
  summarize_by = NULL,
  weighted = TRUE,
  se = FALSE,
  wtname = NULL,
  strataname = NULL
)
}
\arguments{
\item{prepped_dt}{A prepared dataset in data.table format with
the variable to summarize, the variable to summarize by, and the weights,
if used.}

\item{summarize_var}{Name of the variable to summarize. Default is NULL.}

\item{summarize_by}{Name of the variable to summarize the summarize_var by.
Default is NULL.}

\item{weighted}{Whether the data is weighted. Default is TRUE.}

\item{se}{Whether to calculate standard error. Default is FALSE. Will be set
to FALSE if weighted is FALSE.}

\item{wtname}{Name of the weight column to use. Default is NULL. Must be specified
when weighted = TRUE.}

\item{strataname}{Name of strata name to bring in. Default is NULL.}
}
\value{
List of unweighted and weighted numeric summaries including count, min,
max, mean, se, and median.
}
\description{
Summarize a numeric variable
}
\examples{

require(data.table)
require(stringr)
require(dplyr)
require(srvyr)
DT = hts_prep_variable(summarize_var = 'speed_mph',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$num
hts_summary_num(prepped_dt = DT,
                summarize_var = 'speed_mph',
                wtname = 'trip_weight')
DT = hts_prep_variable(summarize_var = 'speed_mph',
                   summarize_by = 'age', 
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$num
hts_summary_num(prepped_dt = DT,
                summarize_var = 'speed_mph',
                summarize_by = 'age',
                wtname = 'trip_weight')
}
