% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variable_list}
\alias{variable_list}
\title{List of variables}
\format{
\subsection{\code{variable_list}}{

A data frame with 55 rows and 10 columns:
\describe{
\item{variable}{Name of the variable}
\item{is_checkbox}{The variable is a 'Select all that Apply' question}
\item{hh}{The variable exists in the hh table}
\item{person}{The variable exists in the person table}
\item{day}{The variable exists in the day table}
\item{trip}{The variable exists in the trip table}
\item{vehicle}{The variable exists in the vehicle table}
\item{data_type}{Data type of the variable ("iteger/categorical", "numeric", "character")}
\item{description}{A description of the variable}
\item{shared_name}{Shared part of name for checkbox variable (e.g., race_1 -> race) or variable name (e.g., age -> age)}
}
}
}
\usage{
variable_list
}
\description{
A dataset containing information about all variables in data. This must contain one binary column indicating table location for each table in the data (ie., hh, person, day, trip, vehicle).
}
\keyword{datasets}
